server/requirements.txt
astapi
uvicorn
sqlalchemy
alembic
pydantic
bcrypt
python-jose
celery
redis
pandas
yt-dlp
